remove unused methods from SortHelper; hide some implementation details
