+++ Forwards.hpp
@@ -67,6 +67,8 @@ typedef List<Literal*> LiteralList;
 typedef Stack<Literal*> LiteralStack;
 typedef VirtualIterator<Literal*> LiteralIterator;
 
+class AtomicSort;
+
 class Inference;
 
 class Unit;
