+++ Kernel/SortHelper.cpp
@@ -29,12 +29,32 @@
 using namespace std;
 using namespace Kernel;
 
+enum CollectWhat {
+  COLLECT_TERM,
+  COLLECT_TERMLIST,
+  COLLECT_SPECIALTERM,
+  COLLECT_FORMULA,
+  BIND,
+  UNBIND,
+};
+
+struct CollectTask {
+  CollectTask(CollectWhat what) : fncTag(what) {}
+  CollectWhat fncTag;
+  union {
+    TermList ts;
+    Term* t; // shared by TERM and SPECIALTERM
+    Formula* f;
+    VList* vars; // to bind/unbind by BIND/UNBIND
+  };
+  TermList contextSort; // only used by TERMLIST and SPECIALTERM
+};
 
 /**
  * Return the type of a term or a literal @c t
  * @author Andrei Voronkov
  */
-OperatorType* SortHelper::getType(Term const* t)
+static OperatorType* getType(Term const* t)
 {
   if (t->isLiteral())
     return env.signature->getPredicate(t->functor())->predType();
@@ -262,21 +282,6 @@ TermList SortHelper::getEqualityArgumentSort(const Literal* lit)
 } //
 
 /**
- * Return sort of term @c trm that appears inside literal @c lit.
- */
-TermList SortHelper::getTermSort(TermList trm, Literal* lit)
-{
-  if (trm.isTerm()) {
-    return getResultSort(trm.term());
-  }
-  if(!trm.isVar()){
-    cout << "ERROR with " << trm.toString() << " in " << lit->toString() << endl;
-  }
-  ASS(trm.isVar());
-  return getVariableSort(trm, lit);
-}
-
-/**
  * Return sort of variable @c var in term or literal @c t
  *
  * Variable @c var must occur in @c t.
@@ -350,7 +355,7 @@ bool SortHelper::tryGetVariableSort(unsigned var, Formula* f, TermList& res)
  * @since 13/02/2017 Vienna
  * @author Martin Suda
  */
-void SortHelper::collectVariableSortsIter(CollectTask task, DHMap<unsigned,TermList>& map, bool ignoreBound)
+static void collectVariableSortsIter(CollectTask task, DHMap<unsigned,TermList>& map, bool ignoreBound = false)
 {
   Stack<CollectTask> todo;
   ZIArray<unsigned> bound;
@@ -362,12 +367,12 @@ void SortHelper::collectVariableSortsIter(CollectTask task, DHMap<unsigned,TermL
     switch(task.fncTag) {
       case COLLECT_TERM: {
         Term* term = task.t;
-    
+
         unsigned position = 0;
         for (TermList* ts = term->args(); ts->isNonEmpty(); ts = ts->next()) {
           CollectTask newTask(COLLECT_TERMLIST);
           newTask.ts = *ts;
-          newTask.contextSort = getArgSort(term, position++);
+          newTask.contextSort = SortHelper::getArgSort(term, position++);
           todo.push(newTask);
         }
