+++ Kernel/TypedTermList.hpp
@@ -11,8 +11,10 @@
 #ifndef __Kernel_TypedTermList__
 #define __Kernel_TypedTermList__
 
-#include "Kernel/SortHelper.hpp"
 #include <tuple>
+
+#include "Term.hpp"
+#include "SortHelper.hpp"
 #include "Lib/Reflection.hpp"
 
 namespace Kernel {
